aws-playground/aws-client/external-capturing/requirements.txt
opencv-python
boto3
cryptography
pyobjc-framework-AVFoundation
